Grouped CLI- and GUI-related code into modules
Some users (and myself) have experienced difficulty compiling TMS
Express on Windows. Separating out the code for the GUI is the first
step toward distributing separate CLI/GUI builds, so Windows users can
at least take advantage of the program while I work on Qt...
