+++ tms_express/ui/gui/controlpanels/ControlPanelLpcView.cpp
@@ -1,17 +1,13 @@
 // Copyright 2023 Joseph Bellahcen <joeclb@icloud.com>
 
-#include "ui/Control_Panels/ControlPanelPitchView.hpp"
+#include "ui/gui/controlpanels/ControlPanelLpcView.hpp"
 
 #include <QCheckBox>
-#include <QFrame>
-#include <QSlider>
 #include <QLabel>
 #include <QLineEdit>
 #include <QWidget>
 
-#include <string>
-
-#include "ui/Control_Panels/ControlPanelView.hpp"
+#include "ui/gui/controlpanels/ControlPanelView.hpp"
 
 namespace tms_express::ui {
 
@@ -19,10 +15,13 @@ namespace tms_express::ui {
 // Initializers ///////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
 
-ControlPanelPitchView::ControlPanelPitchView(QWidget *parent):
-    ControlPanelView("Pitch Analysis", parent) {
+ControlPanelLpcView::ControlPanelLpcView(QWidget *parent):
+    ControlPanelView("LPC Analysis", parent)  {
     //
     // Initialize parameters
+    auto analysis_window_label = new QLabel("Analysis window (ms)", this);
+    analysis_window_line_ = new QLineEdit("25.0", this);
+
     hpf_checkbox_ = new QCheckBox("Highpass filter (Hz)", this);
     hpf_line_ = new QLineEdit("100", this);
 
@@ -32,14 +31,11 @@ ControlPanelPitchView::ControlPanelPitchView(QWidget *parent):
     preemphasis_checkbox_ = new QCheckBox("Pre-emphasis filter (alpha)", this);
     preemphasis_line_ = new QLineEdit("0.9375", this);
 
-    auto maxPitchLabel = new QLabel("Max pitch (Hz)", this);
-    max_pitch_frq_line_ = new QLineEdit("500", this);
-
-    auto minPitchLabel = new QLabel("Min pitch (Hz)", this);
-    min_pitch_frq_line_ = new QLineEdit("50", this);
-
     // Construct layout
     auto row = grid->rowCount();
+
+    grid->addWidget(analysis_window_label, row, 0);
+    grid->addWidget(analysis_window_line_, row++, 1);
 
     grid->addWidget(hpf_checkbox_, row, 0);
     grid->addWidget(hpf_line_, row++, 1);
@@ -48,38 +44,28 @@ ControlPanelPitchView::ControlPanelPitchView(QWidget *parent):
     grid->addWidget(lpf_line_, row++, 1);
 
     grid->addWidget(preemphasis_checkbox_, row, 0);
-    grid->addWidget(preemphasis_line_, row++, 1);
-
-    auto line2 = new QFrame(this);
-    line2->setFrameShape(QFrame::HLine);
-    grid->addWidget(line2, row++, 0, 1, 2);
-
-    grid->addWidget(maxPitchLabel, row, 0);
-    grid->addWidget(max_pitch_frq_line_, row++, 1);
-
-    grid->addWidget(minPitchLabel, row, 0);
-    grid->addWidget(min_pitch_frq_line_, row, 1);
+    grid->addWidget(preemphasis_line_, row, 1);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 // Overloaded Methods /////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
 
-void ControlPanelPitchView::reset() {
+void ControlPanelLpcView::reset() {
     hpf_checkbox_->setChecked(false);
     hpf_line_->setText("100");
 
-    lpf_checkbox_->setChecked(true);
+    lpf_checkbox_->setChecked(false);
     lpf_line_->setText("800");
 
-    preemphasis_checkbox_->setChecked(false);
+    preemphasis_checkbox_->setChecked(true);
     preemphasis_line_->setText("0.9375");
-
-    max_pitch_frq_line_->setText("500");
-    min_pitch_frq_line_->setText("50");
 }
 
-void ControlPanelPitchView::configureSlots() {
+void ControlPanelLpcView::configureSlots() {
+    connect(analysis_window_line_, &QLineEdit::editingFinished, this,
+        &ControlPanelView::stateChanged);
+
     connect(hpf_checkbox_, &QCheckBox::released, this,
         &ControlPanelView::stateChanged);
 
@@ -97,48 +83,38 @@ void ControlPanelPitchView::configureSlots() {
 
     connect(preemphasis_line_, &QLineEdit::editingFinished, this,
         &ControlPanelView::stateChanged);
-
-    connect(max_pitch_frq_line_, &QLineEdit::editingFinished, this,
-        &ControlPanelView::stateChanged);
-
-    connect(min_pitch_frq_line_, &QLineEdit::editingFinished, this,
-        &ControlPanelView::stateChanged);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 // Accessors //////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
 
-bool ControlPanelPitchView::getHpfEnabled() {
+float ControlPanelLpcView::getAnalysisWindowWidth() {
+    return analysis_window_line_->text().toFloat();
+}
+
+bool ControlPanelLpcView::getHpfEnabled() {
     return hpf_checkbox_->isChecked();
 }
 
-int ControlPanelPitchView::getHpfCutoff() {
+int ControlPanelLpcView::getHpfCutoff() {
     return hpf_line_->text().toInt();
 }
 
-bool ControlPanelPitchView::getLpfEnabled() {
+bool ControlPanelLpcView::getLpfEnabled() {
     return lpf_checkbox_->isChecked();
 }
 
-int ControlPanelPitchView::getLpfCutoff() {
+int ControlPanelLpcView::getLpfCutoff() {
     return lpf_line_->text().toInt();
 }
 
-bool ControlPanelPitchView::getPreEmphasisEnabled() {
+bool ControlPanelLpcView::getPreEmphasisEnabled() {
     return preemphasis_checkbox_->isChecked();
 }
 
-float ControlPanelPitchView::getPreEmphasisAlpha() {
+float ControlPanelLpcView::getPreEmphasisAlpha() {
     return preemphasis_line_->text().toFloat();
-}
-
-int ControlPanelPitchView::getMaxPitchFrq() {
-    return max_pitch_frq_line_->text().toInt();
-}
-
-int ControlPanelPitchView::getMinPitchFrq() {
-    return min_pitch_frq_line_->text().toInt();
 }
 
 };  // namespace tms_express::ui
