+++ tms_express/ui/gui/controlpanels/ControlPanelPitchView.cpp
@@ -1,13 +1,17 @@
 // Copyright 2023 Joseph Bellahcen <joeclb@icloud.com>
 
-#include "ui/Control_Panels/ControlPanelLpcView.hpp"
+#include "ui/gui/controlpanels/ControlPanelPitchView.hpp"
 
 #include <QCheckBox>
+#include <QFrame>
+#include <QSlider>
 #include <QLabel>
 #include <QLineEdit>
 #include <QWidget>
 
-#include "ui/Control_Panels/ControlPanelView.hpp"
+#include <string>
+
+#include "ui/gui/controlpanels/ControlPanelView.hpp"
 
 namespace tms_express::ui {
 
@@ -15,13 +19,10 @@ namespace tms_express::ui {
 // Initializers ///////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
 
-ControlPanelLpcView::ControlPanelLpcView(QWidget *parent):
-    ControlPanelView("LPC Analysis", parent)  {
+ControlPanelPitchView::ControlPanelPitchView(QWidget *parent):
+    ControlPanelView("Pitch Analysis", parent) {
     //
     // Initialize parameters
-    auto analysis_window_label = new QLabel("Analysis window (ms)", this);
-    analysis_window_line_ = new QLineEdit("25.0", this);
-
     hpf_checkbox_ = new QCheckBox("Highpass filter (Hz)", this);
     hpf_line_ = new QLineEdit("100", this);
 
@@ -31,11 +32,14 @@ ControlPanelLpcView::ControlPanelLpcView(QWidget *parent):
     preemphasis_checkbox_ = new QCheckBox("Pre-emphasis filter (alpha)", this);
     preemphasis_line_ = new QLineEdit("0.9375", this);
 
+    auto maxPitchLabel = new QLabel("Max pitch (Hz)", this);
+    max_pitch_frq_line_ = new QLineEdit("500", this);
+
+    auto minPitchLabel = new QLabel("Min pitch (Hz)", this);
+    min_pitch_frq_line_ = new QLineEdit("50", this);
+
     // Construct layout
     auto row = grid->rowCount();
-
-    grid->addWidget(analysis_window_label, row, 0);
-    grid->addWidget(analysis_window_line_, row++, 1);
 
     grid->addWidget(hpf_checkbox_, row, 0);
     grid->addWidget(hpf_line_, row++, 1);
@@ -44,28 +48,38 @@ ControlPanelLpcView::ControlPanelLpcView(QWidget *parent):
     grid->addWidget(lpf_line_, row++, 1);
 
     grid->addWidget(preemphasis_checkbox_, row, 0);
-    grid->addWidget(preemphasis_line_, row, 1);
+    grid->addWidget(preemphasis_line_, row++, 1);
+
+    auto line2 = new QFrame(this);
+    line2->setFrameShape(QFrame::HLine);
+    grid->addWidget(line2, row++, 0, 1, 2);
+
+    grid->addWidget(maxPitchLabel, row, 0);
+    grid->addWidget(max_pitch_frq_line_, row++, 1);
+
+    grid->addWidget(minPitchLabel, row, 0);
+    grid->addWidget(min_pitch_frq_line_, row, 1);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 // Overloaded Methods /////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
 
-void ControlPanelLpcView::reset() {
+void ControlPanelPitchView::reset() {
     hpf_checkbox_->setChecked(false);
     hpf_line_->setText("100");
 
-    lpf_checkbox_->setChecked(false);
+    lpf_checkbox_->setChecked(true);
     lpf_line_->setText("800");
 
-    preemphasis_checkbox_->setChecked(true);
+    preemphasis_checkbox_->setChecked(false);
     preemphasis_line_->setText("0.9375");
+
+    max_pitch_frq_line_->setText("500");
+    min_pitch_frq_line_->setText("50");
 }
 
-void ControlPanelLpcView::configureSlots() {
-    connect(analysis_window_line_, &QLineEdit::editingFinished, this,
-        &ControlPanelView::stateChanged);
-
+void ControlPanelPitchView::configureSlots() {
     connect(hpf_checkbox_, &QCheckBox::released, this,
         &ControlPanelView::stateChanged);
 
@@ -83,38 +97,48 @@ void ControlPanelLpcView::configureSlots() {
 
     connect(preemphasis_line_, &QLineEdit::editingFinished, this,
         &ControlPanelView::stateChanged);
+
+    connect(max_pitch_frq_line_, &QLineEdit::editingFinished, this,
+        &ControlPanelView::stateChanged);
+
+    connect(min_pitch_frq_line_, &QLineEdit::editingFinished, this,
+        &ControlPanelView::stateChanged);
 }
 
 ///////////////////////////////////////////////////////////////////////////////
 // Accessors //////////////////////////////////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////
 
-float ControlPanelLpcView::getAnalysisWindowWidth() {
-    return analysis_window_line_->text().toFloat();
-}
-
-bool ControlPanelLpcView::getHpfEnabled() {
+bool ControlPanelPitchView::getHpfEnabled() {
     return hpf_checkbox_->isChecked();
 }
 
-int ControlPanelLpcView::getHpfCutoff() {
+int ControlPanelPitchView::getHpfCutoff() {
     return hpf_line_->text().toInt();
 }
 
-bool ControlPanelLpcView::getLpfEnabled() {
+bool ControlPanelPitchView::getLpfEnabled() {
     return lpf_checkbox_->isChecked();
 }
 
-int ControlPanelLpcView::getLpfCutoff() {
+int ControlPanelPitchView::getLpfCutoff() {
     return lpf_line_->text().toInt();
 }
 
-bool ControlPanelLpcView::getPreEmphasisEnabled() {
+bool ControlPanelPitchView::getPreEmphasisEnabled() {
     return preemphasis_checkbox_->isChecked();
 }
 
-float ControlPanelLpcView::getPreEmphasisAlpha() {
+float ControlPanelPitchView::getPreEmphasisAlpha() {
     return preemphasis_line_->text().toFloat();
+}
+
+int ControlPanelPitchView::getMaxPitchFrq() {
+    return max_pitch_frq_line_->text().toInt();
+}
+
+int ControlPanelPitchView::getMinPitchFrq() {
+    return min_pitch_frq_line_->text().toInt();
 }
 
 };  // namespace tms_express::ui
